+++ protected/modules/users/controllers/TeacherDetailsController.php
@@ -29,7 +29,7 @@ class TeacherDetailsController extends Controller
 	{
 		return array(
 			array('allow',  // allow all users to perform 'index' and 'view' actions
-				  'actions'=>array('index','view','create','update','admin','delete', 'upload', 'deleteUpload'),
+				  'actions'=>array('index','view','create','update','admin','delete', 'upload', 'deleteUpload', 'uploadFile', 'deleteUploadFile'),
 				  'roles'=>array('admin'),
 			),
 			array('deny',  // deny all users
@@ -59,12 +59,17 @@ class TeacherDetailsController extends Controller
 		$model=$this->loadModel($id);
 
 		$tmpDIR = Yii::getPathOfAlias("webroot") . '/uploads/temp/';
-		$tmpUrl = Yii::app()->createAbsoluteUrl('/uploads/temp/');
+		$tmpUrl = Yii::app()->baseUrl .'/uploads/temp/';
 		$avatarDIR = Yii::getPathOfAlias("webroot") . "/uploads/teachers/";
-		$avatarUrl = Yii::app()->createAbsoluteUrl('/uploads/teachers/');
+		$avatarUrl = Yii::app()->baseUrl .'/uploads/teachers/';
+		$fileDIR = Yii::getPathOfAlias("webroot") . "/uploads/teachers/files/";
+		$fileUrl = Yii::app()->baseUrl .'/uploads/teachers/files/';
+
+		if(!is_dir($fileDIR))
+			mkdir($fileDIR);
 
 		$avatar = array();
-		$flag = true;
+		$flag = false;
 		if ($model->avatar and file_exists($avatarDIR.$model->avatar)) {
 			$file = $model->avatar;
 			$avatar = array(
@@ -73,7 +78,18 @@ class TeacherDetailsController extends Controller
 				'size' => filesize($avatarDIR . $file),
 				'serverName' => $file,
 			);
-			$flag = false;
+		}
+
+		$flag2 = false;
+		$resumeFile = array();
+		if ($model->file and file_exists($fileDIR.$model->file)) {
+			$file = $model->file;
+			$resumeFile = array(
+					'name' => $file,
+					'src' => $fileUrl . '/' . $file,
+					'size' => filesize($fileDIR . $file),
+					'serverName' => $file,
+			);
 		}
 		if(isset($_POST['TeacherDetails']))
 		{
@@ -99,12 +115,25 @@ class TeacherDetailsController extends Controller
 				);
 				$flag = true;
 			}
+			if (isset($_POST['TeacherDetails']['file'])&& file_exists($tmpDIR.$_POST['TeacherDetails']['file'])) {
+				$file = $_POST['TeacherDetails']['file'];
+				$resumeFile = array(
+						'name' => $file,
+						'src' => $tmpUrl . '/' . $file,
+						'size' => filesize($tmpDIR . $file),
+						'serverName' => $file,
+				);
+				$flag2 = true;
+			}
 			if($model->save())
 			{
 				if ($flag && $model->avatar && file_exists($tmpDIR.$model->avatar)) {
 					$imager = new Imager();
 					$imager->resize($tmpDIR.$model->avatar, $avatarDIR.$model->avatar, 400, 400);
 					unlink($tmpDIR.$model->avatar);
+				}
+				if ($flag2 && $model->file and file_exists($tmpDIR.$model->file)) {
+					rename($tmpDIR.$model->file,$fileDIR.$model->file);
 				}
 				Yii::app()->user->setFlash('success' ,'<span class="icon-check"></span>&nbsp;&nbsp;اطلاعات با موفقیت ذخیره شد.');
 				if(isset($_GET['return']) && $_GET['return'] == true)
@@ -116,8 +145,9 @@ class TeacherDetailsController extends Controller
 		}
 
 		$this->render('update',array(
-				'model'=>$model,
-				'avatar' => $avatar
+			'model'=>$model,
+			'avatar' => $avatar,
+			'file' => $resumeFile
 		));
 	}
 
@@ -129,9 +159,12 @@ class TeacherDetailsController extends Controller
 	public function actionDelete($id)
 	{
 		$avatarDIR = Yii::getPathOfAlias("webroot") . "/uploads/teachers/";
+		$fileDIR = Yii::getPathOfAlias("webroot") . "/uploads/teachers/files/";
 		$model = $this->loadModel($id);
-		if(file_exists($avatarDIR.$model->avatar))
+		if($model->avatar && file_exists($avatarDIR.$model->avatar))
 			unlink($avatarDIR.$model->avatar);
+		if($model->file && file_exists($fileDIR.$model->file))
+			unlink($fileDIR.$model->file);
 		$model->delete();
 
 		// if AJAX request (triggered by deletion via admin grid view), we should not redirect the browser
@@ -226,7 +259,59 @@ class TeacherDetailsController extends Controller
 			$model = TeacherDetails::model()->findByAttributes(array('avatar' => $fileName));
 			if ($model) {
 				if (@unlink($Dir . $model->avatar)) {
-					$model->updateByPk($model->id, array('avatar' => null));
+					$model->updateByPk($model->user_id, array('avatar' => null));
+					$response = ['state' => 'ok', 'msg' => $this->implodeErrors($model)];
+				} else
+					$response = ['state' => 'error', 'msg' => 'مشکل ایجاد شده است'];
+			} else {
+				@unlink($tempDir . $fileName);
+				$response = ['state' => 'ok', 'msg' => 'حذف شد.'];
+			}
+			echo CJSON::encode($response);
+			Yii::app()->end();
+		}
+	}
+
+	/**
+	 * file uploader
+	 */
+	public function actionUploadFile()
+	{
+		$tempDir = Yii::getPathOfAlias("webroot") . '/uploads/temp';
+
+		if (!is_dir($tempDir))
+			mkdir($tempDir);
+		if (isset($_FILES)) {
+			$file = $_FILES['file'];
+			$ext = pathinfo($file['name'], PATHINFO_EXTENSION);
+			$file['name'] = Controller::generateRandomString(5) . time();
+			while (file_exists($tempDir . DIRECTORY_SEPARATOR . $file['name']))
+				$file['name'] = Controller::generateRandomString(5) . time();
+			$file['name'] = $file['name'] . '.' . $ext;
+			if (move_uploaded_file($file['tmp_name'], $tempDir . DIRECTORY_SEPARATOR . CHtml::encode($file['name'])))
+				$response = ['state' => 'ok', 'fileName' => CHtml::encode($file['name'])];
+			else
+				$response = ['state' => 'error', 'msg' => 'فایل آپلود نشد.'];
+		} else
+			$response = ['state' => 'error', 'msg' => 'فایلی ارسال نشده است.'];
+		echo CJSON::encode($response);
+		Yii::app()->end();
+	}
+
+	public function actionDeleteUploadFile()
+	{
+		$Dir = Yii::getPathOfAlias("webroot") . '/uploads/teachers/files/';
+
+		if (isset($_POST['fileName'])) {
+
+			$fileName = $_POST['fileName'];
+
+			$tempDir = Yii::getPathOfAlias("webroot") . '/uploads/temp/';
+
+			$model = TeacherDetails::model()->findByAttributes(array('file' => $fileName));
+			if ($model) {
+				if (@unlink($Dir . $model->file)) {
+					$model->updateByPk($model->user_id, array('file' => null));
 					$response = ['state' => 'ok', 'msg' => $this->implodeErrors($model)];
 				} else
 					$response = ['state' => 'error', 'msg' => 'مشکل ایجاد شده است'];
